FinsmartAI/FinSmartBackend/shiva_repo/FinSmartAI/pyproject.toml
[project]
name = "finsmartai"
version = "0.1.0"
description = "Professional stock analysis using AI agents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.86.0",
    "crewai-tools>=0.15.0",
    "langchain>=0.3.13",
    "langchain-community>=0.3.13",
    "yfinance>=0.2.51",
    "requests>=2.32.3",
    "python-dotenv>=1.0.1",
    "fastapi>=0.128.8",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.5",
    "httpx>=0.28.1",
    "openai>=1.59.7",
    "python-multipart>=0.0.20",
    "annotated-doc>=0.0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["Agent", "api"]
